TODO: Figure out user information retrieval
calls for the shelter_comments component to the backend is generally set up, we still need to figure out how to do user retrieval. Ideally, for the shelter_comments we do the same but only twice, once per sender and once per reciever. we still need some way to retrieve user information though